Add a AbstractJdbcSourceChunkSplitter to handle common logic
